backend/requirements.txt
beautifulsoup4>=4.14.3
docling>=2.69.1
feedparser>=6.0.12
markdown>=3.9
openai>=2.33.0
psycopg2-binary>=2.9.12
pydantic>=2.13.3
email-validator>=2.1.0
python-dotenv>=1.2.1
requests>=2.32.5
sqlalchemy>=2.0.49
youtube-transcript-api>=1.2.4
fastapi>=0.115.0
uvicorn>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt,argon2]>=1.7.4
argon2-cffi>=23.1.0
fastapi-mail>=1.4.1
ruff>=0.4.0
mypy>=1.10.0
